+++ Frontend/package.json
@@ -16,12 +16,12 @@
     "react-bootstrap": "^2.10.5",
     "react-dom": "^18.3.1",
     "react-redux": "^9.1.2",
-    "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
+    "start": "vite",
+    "build": "vite build",
+    "serve": "vite preview",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
