+++ Frontend/vite.config.js
@@ -1,0 +1,9 @@
+import { defineConfig } from "vite";
+import react from "@vitejs/plugin-react";
+
+export default defineConfig({
+  plugins: [react()],
+  build: {
+    outDir: "build", // Directorio de salida similar a CRA
+  },
+});
